Module redux::bitio

Bit level I/O operations.

Structs

BitReader
A class for wrapping a byte input stream in a bit based interface.

BitWriter
A class for wrapping a byte output stream in a bit based interface.

Traits

BitRead
A trait for object that allow reading one bit at a time.

BitWrite
A trait for object that allow writing one bit at a time.

ByteCount
A trait for counting the number of bytes flowing trough a Read or Write implementation.

ByteRead
A trait for object that allow reading one byte at a time.

ByteWrite
A trait for object that allow writing one byte at a time.