Struct redux::codec::Codec

pub struct Codec {
    // some fields omitted
}

The current state of the encoder and decoder.

Methods

impl Codec

fn new(m: Box<Model>) -> Codec

Creates and initializes the codec for encoding or decoding.

fn compress_symbol(&mut self, symbol: usize, output: &mut BitWrite) -> Result<()>

Compresses a symbol and outputs some bits depending on the state of the codec.

fn compress_bytes(&mut self, input: &mut ByteRead, output: &mut BitWrite) -> Result<()>

Compresses an entire byte stream outputting the EOF symbol and all bits for unambigous encoding.

fn decompress_symbol(&mut self, input: &mut BitRead) -> Result<usize>

Decompresses a symbol reading some bits until the symbol can be decoded.

fn decompress_bytes(&mut self, input: &mut BitRead, output: &mut ByteWrite) -> Result<()>

Decompresses a whole bit stream until the EOF symbol is found.