Module redux::model

Symbol frequency distribution models.

Structs

AdaptiveLinearModel
Adaptive model that uses a simple array for cumulative freq and simple, but slow linear algorithms for operations.

AdaptiveTreeModel
Adaptive model that uses a Binary Indexed Tree for storing cumulative frequencies.

Parameters
Model parameters that specifies the common property of the models.

Traits

Model
Trait for the probability models behind arithmetic coding. Possible implementations may include static models with fixed probabilities or and adaptive model that continuously updates cumulative frequencies.