Struct redux::model::AdaptiveLinearModel

pub struct AdaptiveLinearModel { /* fields omitted */ }

Adaptive model that uses a simple array for cumulative freq and simple, but slow linear algorithms for operations.

Methods

impl AdaptiveLinearModel

fn new(p: Parameters) -> Box<Model>

Initializes the model with the given parameters.