Struct redux::model::Parameters

pub struct Parameters {
    pub symbol_bits: usize,
    pub symbol_eof: usize,
    pub symbol_count: usize,
    pub freq_bits: usize,
    pub freq_max: u64,
    pub code_bits: usize,
    pub code_min: u64,
    pub code_one_fourth: u64,
    pub code_half: u64,
    pub code_three_fourths: u64,
    pub code_max: u64,
}

Model parameters that specifies the common property of the models.

Fields

symbol_bits

Bit width of the symbols being encoded. Usually 8 for byte oriented inputs.

symbol_eof

Code for the EOF symbol. This is the next symbol code after the valid symbols to encode.

symbol_count

Number of possible symbols including the EOF symbol.

freq_bits

Number of bits representing symbol frequencies.

freq_max

Maximum cumulated frequency value for symbols.

code_bits

Number of bits representing the current code ranges.

code_min

Minimum value for code range. This is always zero.

code_one_fourth

Delimiter for the one fourth of the valid code range.

code_half

Delimiter for the half of the valid code range.

code_three_fourths

Delimiter for the three fourths of the valid code range.

code_max

Upper limit of the valid code range.

Methods

impl Parameters

fn new(sym: usize, freq: usize, code: usize) -> Result<Parameters>

Calculates all parameter values based on the symbol, frequency and code width.