Crate redux

Re-exports

pub use callback::paste;

Macros

_callback

callback

Creates a callback instance. Must accept a single argument, so () should be used when no arguments are needed and tuples where more than one value need to be passed.

Structs

ActionId

Time in nanoseconds from std::time::UNIX_EPOCH.

ActionMeta

Action with additional metadata like: id.

ActionWithMeta

Action with additional metadata like: id.

AnyAction

Callback

Dispatcher

Instant

Store

Main struct for the state machine.

SystemTime

A measurement of the system clock, useful for talking to external entities like the file system or other processes.

Timestamp

Time in nanoseconds from std::time::UNIX_EPOCH.

Traits

EnablingCondition

Service

SubStore

Useful when state machine is split into multiple crates. Using this trait we can pass Store<GlobalState, Service, GlobalAction> almost as if it were Store<SubState, Service, SubAction>.

TimeService

Time service.

Type Aliases

Effects

RecursionDepth

Reducer

Function signature for a reducer.