Crate redux

Structs

ActionId

Time in nanoseconds from std::time::UNIX_EPOCH.

ActionMeta

Action with additional metadata like: id.

ActionWithMeta

Action with additional metadata like: id.

Instant

A measurement of a monotonically nondecreasing clock. Opaque and useful only with Duration.

Store

Main struct for the state machine.

SystemTime

A measurement of the system clock, useful for talking to external entities like the file system or other processes.

Timestamp

Time in nanoseconds from std::time::UNIX_EPOCH.

Traits

EnablingCondition

Service

SubStore

Useful when state machine is split into multiple crates. Using this trait we can pass Store<GlobalState, Service, GlobalAction> almost as if it were Store<SubState, Service, SubAction>.

TimeService

Time service.

Functions

monotonic_to_time

Converts monotonic time to nanoseconds since Unix epoch.

Type Aliases

Effects

RecursionDepth

Reducer

Function signature for a reducer.